requirements.txt
streamlit
fastapi
uvicorn
google-generativeai
sqlalchemy
passlib
bcrypt
python-multipart
pillow
requests
